requirements.txt
python-telegram-bot[ext]==21.6
requests==2.32.3
flask==3.0.3
# python dotenv
python-dotenv
